+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Image resources: Support `res://`, `www` and `shared://` for [attachments](README.md#property-attachments) and [attachments.personIcon](README.md#property-personicon)
 ### iOS
 - Since iOS 14: Show a notification always in the notification center like on Android, also if [iOSForeground](#property-iosforeground) is `false`.
+- Renamed internal plugin method `check` to `hasPermission` to make it consistent with `cordova.plugins.notification.local.hasPermission`
 
 ## Version 1.1.0 (23.12.2024)